requirements.txt
wmi==1.5.1
pywin32==305
colorama==0.4.6
pillow==9.4.0
infi.systray==0.1.7
